Restructure for documentation
