+++ README.md
@@ -12,6 +12,36 @@
 This repo contains a set of CLI scripts leveraged throughout the Garden
 repositories for common tasks and CI automation.
 
+## Installation
+
+```sh
+npm install --save-dev @zendeskgarden/tool-shed
+```
+
+... or for one-off execution
+
+```sh
+npx -p @zendeskgarden/tool-shed garden [command]
+```
+
+## Usage
+
+All commands are available via the `garden` executable. Commands may be
+chained so that the output from one command provides input to another. Each
+command also provides a functional equivalent for programmatic usage. See
+the individual command README for details:
+
+- [`garden github-branch`](src/github/branch)
+- [`garden github-commit`](src/github/commit)
+- [`garden github-deploy`](src/github/deploy)
+- [`garden github-owner`](src/github/owner)
+- [`garden github-pages`](src/github/pages)
+- [`garden github-repo`](src/github/repo)
+- [`garden github-token`](src/github/token)
+- [`garden netlify-deploy`](src/netlify/deploy)
+- [`garden netlify-site-id`](src/netlify/site_id)
+- [`garden netlify-token`](src/netlify/token)
+
 ## Contribution
 
 Thanks for your interest in Garden! Community involvement helps make our
