+++ package.json
@@ -28,11 +28,10 @@
     "start": "bin/index.js",
     "tag": "[ `git rev-parse --abbrev-ref HEAD` = 'master' ] && standard-version --no-verify",
     "test": "prettier-package-json --list-different && yarn format:all --check && yarn lint",
-    "watch": "yarn tsc --watch"
+    "watch": "yarn build --watch"
   },
   "dependencies": {
     "@octokit/rest": "17.1.4",
-    "@zendeskgarden/react-theming": "8.4.1",
     "chalk": "4.0.0",
     "commander": "5.0.0",
     "envalid": "6.0.1",
@@ -59,9 +58,7 @@
     "markdownlint-cli": "0.22.0",
     "prettier": "2.0.2",
     "prettier-package-json": "2.1.3",
-    "react": "16.13.1",
     "standard-version": "7.1.0",
-    "styled-components": "5.0.1",
     "typescript": "3.8.3"
   },
   "keywords": [
