+++ src/github/owner/index.ts
@@ -5,43 +5,44 @@
  * found at http://www.apache.org/licenses/LICENSE-2.0.
  */
 
-import { handleErrorMessage, handleSuccessMessage } from '../utils';
+import { handleErrorMessage, handleSuccessMessage } from '../../utils';
 import { Command } from 'commander';
 import { Ora } from 'ora';
 import execa from 'execa';
 
 /**
- * Execute the `github-repo` command.
+ * Execute the `github-owner` command.
  *
  * @param {string} [path] Path to a git directory.
  * @param {Ora} [spinner] Terminal spinner.
  *
- * @returns {Promise<string>} The repo name provided by a CI environment
- * variable or extracted from the given git repository.
+ * @returns {Promise<string>} The owner provided by a CI environment variable
+ * or extracted from the given git repository.
  */
 export const execute = async (path?: string, spinner?: Ora): Promise<string | undefined> => {
   let retVal: string | undefined;
 
   if (process.env.TRAVIS_REPO_SLUG) {
-    retVal = process.env.TRAVIS_REPO_SLUG.split('/')[1];
+    retVal = process.env.TRAVIS_REPO_SLUG[0];
   } else {
-    retVal = process.env.CIRCLE_PROJECT_REPONAME as string;
+    retVal = process.env.CIRCLE_PROJECT_USERNAME as string;
   }
 
   if (!retVal) {
-    const args = ['rev-parse', '--show-toplevel'];
+    const args = ['ls-remote', '--get-url'];
 
     if (path) {
       args.unshift('-C', path);
     }
 
     try {
-      const workingTree = await execa('git', args);
-      const repo = await execa('basename', [workingTree.stdout]);
+      const remote = await execa('git', args);
+      const url = await execa('dirname', [remote.stdout]);
+      const owner = await execa('basename', [url.stdout]);
 
-      retVal = repo.stdout.toString();
+      retVal = owner.stdout.toString();
     } catch (error) {
-      handleErrorMessage(error, 'github-repo', spinner);
+      handleErrorMessage(error, 'github-owner', spinner);
     }
   }
 
@@ -49,21 +50,21 @@ export const execute = async (path?: string, spinner?: Ora): Promise<string | un
 };
 
 export default (spinner: Ora) => {
-  const command = new Command('github-repo');
+  const command = new Command('github-owner');
 
   return command
-    .description('show GitHub repo name')
+    .description('show GitHub repository owner')
     .arguments('[path]')
     .action(async path => {
       try {
         spinner.start();
 
-        const repo = await execute(path, spinner);
+        const owner = await execute(path, spinner);
 
-        if (repo) {
-          handleSuccessMessage(repo, spinner);
+        if (owner) {
+          handleSuccessMessage(owner, spinner);
         } else {
-          spinner.fail('GitHub repo not found');
+          spinner.fail('GitHub owner not found');
           process.exit(1);
         }
       } finally {
