+++ src/netlify/index.ts
@@ -6,5 +6,5 @@
  */
 
 export { default as deployCommand, execute as deploy } from './deploy';
-export { default as siteIdCommand, execute as siteId } from './siteId';
+export { default as siteIdCommand, execute as siteId } from './site_id';
 export { default as tokenCommand, execute as token } from './token';
